1.13.1 worlds not loading due to biomes being an int array

+++ src/main/java/net/querz/mcaselector/version/anvil113/Anvil113ChunkDataProcessor.java
@@ -21,7 +21,7 @@ public class Anvil113ChunkDataProcessor implements ChunkDataProcessor {
 			zLoop:
 			for (int cz = 0; cz < Tile.CHUNK_SIZE; cz++) {
 
-				byte[] biomes = root.getCompoundTag("Level").getByteArray("Biomes");
+				int[] biomes = root.getCompoundTag("Level").getIntArray("Biomes");
 				int biome = -1;
 				if (biomes.length != 0) {
 					biome = biomes[getIndex(cx, 0, cz)];
